+++ packages/app-page-builder/src/admin/components/BulkActions/SecureActionDelete.tsx
@@ -3,26 +3,27 @@ import { useFolders } from "@webiny/app-aco";
 import { observer } from "mobx-react-lite";
 import { PageListConfig } from "~/admin/config/pages";
 import { usePagesPermissions } from "~/hooks/permissions";
-import { ActionPublish as ActionPublishBase } from "~/admin/components/BulkActions";
+import { ActionDelete as ActionDeleteBase } from "~/admin/components/BulkActions";
 
-export const ActionPublish = observer(() => {
-    const { canPublish } = usePagesPermissions();
-
+export const SecureActionDelete = observer(() => {
+    const { canDelete } = usePagesPermissions();
     const { folderLevelPermissions: flp } = useFolders();
 
     const { useWorker } = PageListConfig.Browser.BulkAction;
     const worker = useWorker();
 
-    const canPublishAll = useMemo(() => {
+    const canDeleteAll = useMemo(() => {
         return worker.items.every(item => {
-            return canPublish() && flp.canManageContent(item.location?.folderId);
+            return (
+                canDelete(item.data.createdBy.id) && flp.canManageContent(item.location?.folderId)
+            );
         });
     }, [worker.items]);
 
-    if (!canPublishAll) {
-        console.log("You don't have permissions to publish pages.");
+    if (!canDeleteAll) {
+        console.log("You don't have permissions to delete pages.");
         return null;
     }
 
-    return <ActionPublishBase />;
+    return <ActionDeleteBase />;
 });
