+++ packages/app-page-builder/src/admin/components/BulkActions/SecureActionPublish.tsx
@@ -3,27 +3,26 @@ import { useFolders } from "@webiny/app-aco";
 import { observer } from "mobx-react-lite";
 import { PageListConfig } from "~/admin/config/pages";
 import { usePagesPermissions } from "~/hooks/permissions";
-import { ActionDelete as ActionDeleteBase } from "~/admin/components/BulkActions";
+import { ActionPublish as ActionPublishBase } from "~/admin/components/BulkActions";
 
-export const ActionDelete = observer(() => {
-    const { canDelete } = usePagesPermissions();
+export const SecureActionPublish = observer(() => {
+    const { canPublish } = usePagesPermissions();
+
     const { folderLevelPermissions: flp } = useFolders();
 
     const { useWorker } = PageListConfig.Browser.BulkAction;
     const worker = useWorker();
 
-    const canDeleteAll = useMemo(() => {
+    const canPublishAll = useMemo(() => {
         return worker.items.every(item => {
-            return (
-                canDelete(item.data.createdBy.id) && flp.canManageContent(item.location?.folderId)
-            );
+            return canPublish() && flp.canManageContent(item.location?.folderId);
         });
     }, [worker.items]);
 
-    if (!canDeleteAll) {
-        console.log("You don't have permissions to delete pages.");
+    if (!canPublishAll) {
+        console.log("You don't have permissions to publish pages.");
         return null;
     }
 
-    return <ActionDeleteBase />;
+    return <ActionPublishBase />;
 });
