--- packages/app-page-builder/src/admin/views/Pages/BulkActions/ActionExport.tsx
@@ -1,1 +1,0 @@
-export { ActionExport } from "~/admin/components/BulkActions";
