--- packages/app-page-builder/src/admin/views/Pages/BulkActions/index.tsx
@@ -1,5 +1,0 @@
-export { ActionDelete } from "./ActionDelete";
-export { ActionExport } from "./ActionExport";
-export { ActionMove } from "./ActionMove";
-export { ActionPublish } from "./ActionPublish";
-export { ActionUnpublish } from "./ActionUnpublish";
